feat: create worfbotdatabase file
